update
update

+++ application/console/traits/Admin.php
@@ -211,7 +211,6 @@ trait Admin
      */
     protected function _renew()
     {
-
         if ($this->app->request->isPost()) {
             $data  = $this->app->request->param();
             $scene = 'add'; // 验证场景，默认是新增
@@ -221,8 +220,10 @@ trait Admin
                     if ($data['cover']) {
                         $data['cover'] = $this->articleCover($data['cover']);
                     } else {
-                        unset($data['cover']);
+                        // unset($data['cover']);
                     }
+                    // 对 ueditor 内容数据的处理
+                    $data['content'] = ueditor_handle($data['content'], $data['title']);
                     # code...
                     break;
 
@@ -230,6 +231,8 @@ trait Admin
                     # code...
                     break;
             }
+
+            // return $data;
 
             // 判断是新增还是更新，如果有键值就是更新，如果没有键值就是新增
             if ($this->app->request->has($this->pk)) {
@@ -317,25 +320,27 @@ trait Admin
                         cache('document_model_list', null);
                         break;
                     case 'Article': // 内容管理时的数据处理
-                        $cover_temp  = $data['cover'];
-                        $temp_arr    = explode('/', $cover_temp);
-                        $images_file = './data/images/';
-                        $time_file   = $temp_arr[0] . '/';
-                        $filename    = $temp_arr[1];
+                        if ($this->app->request->param('cover')) {
+                            $cover_temp  = $data['cover'];
+                            $temp_arr    = explode('/', $cover_temp);
+                            $images_file = './data/images/';
+                            $time_file   = $temp_arr[0] . '/';
+                            $filename    = $temp_arr[1];
 
-                        if (!file_exists($images_file . $time_file)) {
-                            //检查是否有该文件夹，如果没有就创建，并给予最高权限
-                            make_dir($images_file . $time_file);
-                        }
+                            if (!file_exists($images_file . $time_file)) {
+                                //检查是否有该文件夹，如果没有就创建，并给予最高权限
+                                make_dir($images_file . $time_file);
+                            }
 
-                        $temp_file = '../data/temp/' . $cover_temp;
-                        $image     = \think\Image::open($temp_file);
-                        if ($image) {
-                            // 按照原图的比例生成一个最大为150*150的缩略图并保存为thumb.png
-                            $new_file = $images_file . $time_file . $temp_arr[1];
-                            if ($image->thumb(150, 150)->save($new_file)) {
-                                unlink($temp_file);
-                            };
+                            $temp_file = '../data/temp/' . $cover_temp;
+                            $image     = \think\Image::open($temp_file);
+                            if ($image) {
+                                // 按照原图的比例生成一个最大为150*150的缩略图并保存为thumb.png
+                                $new_file = $images_file . $time_file . $temp_arr[1];
+                                if ($image->thumb(150, 150)->save($new_file)) {
+                                    unlink($temp_file);
+                                };
+                            }
                         }
                         // return $data;
                         break;
